requirements.txt
requests==2.32.3
openai==1.52.2
python-dotenv==1.0.1
beautifulsoup4==4.12.3
